Moved support code underneath lib.
