+++ provision-cluster/lib/gke.js
@@ -2,7 +2,7 @@
 
 const container = require('@google-cloud/container')
 const crypto = require('crypto')
-const utils = require('./lib/utils.js')
+const utils = require('./utils.js')
 
 const STATUS_ENUM = container.protos.google.container.v1.Operation.Status
