+++ provision-cluster/lib/gke.test.js
