+++ provision-cluster/lib/mock.js
